# Improve performance of many operations, especially for large binaries