--- a/src/Directory.Packages.props
+++ b/src/Directory.Packages.props
@@ -7,16 +7,16 @@
     <PackageVersion Include="Castle.Windsor" Version="6.0.0" />
     <PackageVersion Include="ClosedXML" Version="0.102.2" />
     <PackageVersion Include="DiffPlex.Wpf" Version="1.4.0" />
-    <PackageVersion Include="DotNet.ReproducibleBuilds" Version="1.1.1"/>
-    <PackageVersion Include="DotNet.ReproducibleBuilds.Isolated" Version="1.1.1"/>
+    <PackageVersion Include="DotNet.ReproducibleBuilds" Version="1.2.25" />
+    <PackageVersion Include="DotNet.ReproducibleBuilds.Isolated" Version="1.2.25" />
     <PackageVersion Include="Microsoft.Debugging.DataModel.DbgModelApiXtn" Version="20230731.1609.0" />
     <PackageVersion Include="Microsoft.Debugging.Platform.DbgX" Version="20240216.1.0" />
-    <PackageVersion Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.2" />
-    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
+    <PackageVersion Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.10" />
+    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.11.1" />
     <PackageVersion Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.77" />
     <PackageVersion Include="Moq" Version="4.18.4" />
-    <PackageVersion Include="MSTest.TestAdapter" Version="3.2.2" />
-    <PackageVersion Include="MSTest.TestFramework" Version="3.2.2" />
+    <PackageVersion Include="MSTest.TestAdapter" Version="3.6.1" />
+    <PackageVersion Include="MSTest.TestFramework" Version="3.6.1" />
     <PackageVersion Include="Nito.AsyncEx" Version="5.1.2" />
   </ItemGroup>
 </Project>