--- a/src/SizeBench.AnalysisEngine.RealPETests/Single Binary/PGOTests.cs
+++ b/src/SizeBench.AnalysisEngine.RealPETests/Single Binary/PGOTests.cs
@@ -540,16 +540,4 @@ public sealed class PGOTests
         Assert.AreEqual("mincore", placement.Lib!.ShortName);
         Assert.IsNull(placement.SourceFile);
     }
-
-    public static IEnumerable<object[]> DynamicDataSourceForSymbolSourcesSupportedTests_Slimmed => 
-        SymbolSourcesSupportedCommonTests.DynamicDataSourceForSymbolSourcesSupportedTests_Slimmed;
-
-    [TestMethod]
-    [DynamicData(nameof(DynamicDataSourceForSymbolSourcesSupportedTests_Slimmed))]
-    public Task SymbolSourcesSupportedWorks(SymbolSourcesSupported symbolSources) =>
-        SymbolSourcesSupportedCommonTests.VerifyNoUnexpectedSymbolTypesCanBeMaterialized(
-            Path.Combine(this.TestContext!.DeploymentDirectory!, "Microsoft.UI.Xaml.dll"),
-            Path.Combine(this.TestContext!.DeploymentDirectory!, "Microsoft.UI.Xaml.pdb"),
-            symbolSources,
-            this.TestContext.CancellationTokenSource.Token);
 }