--- a/src/global.json
+++ b/src/global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "6.0",
+    "version": "8.0",
     "allowPrerelease": false,
     "rollForward": "latestMinor"
   }