… docs/Analysis Engine internals.md
… docs/Opening a Session.md
… docs/Principles.md
… docs/Solution Architecture.md
… docs/Test Strategy.md
… src/SizeBench.AnalysisEngine.RealPETests/Single Binary/UDT_LoadFunctionsTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionDataCacheTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Single Binary/SessionDataCacheTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Single Binary/TypeLayoutItemTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Single Binary/WastefulVirtualItemTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Symbols/CanLoadLayoutTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Symbols/ComplexFunctionSymbolTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Symbols/FunctionCodeFormattedNameTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Symbols/InlineSiteSymbolTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Symbols/SimpleFunctionSymbolTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Symbols/StaticDataSymbolTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionManagedObjects/Symbols/UserDefinedTypeSymbolTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionTasks/EnumerateWastefulVirtualsSessionTaskTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionTasks/LoadTypeLayoutSessionTaskTests.cs
… src/SizeBench.AnalysisEngine.Tests/SessionTasks/LookupSymbolPlacementInBinarySessionTaskTests.cs
… src/SizeBench.AnalysisEngine.Tests/SymbolNameHelperTests.cs
… src/SizeBench.AnalysisEngine/DIAInterop/DIAAdapter.cs
… src/SizeBench.AnalysisEngine/DIAInterop/DiaChunkMarshaling.cs
… src/SizeBench.AnalysisEngine/DIAInterop/IDIAAdapter.cs
… src/SizeBench.AnalysisEngine/DIAInterop/IDiaEnumLineNumbersHandCoded.cs
… src/SizeBench.AnalysisEngine/DIAInterop/IDiaEnumSectionContribsHandCoded.cs
… src/SizeBench.AnalysisEngine/DIAInterop/IDiaEnumSymbolsByAddr2HandCoded.cs
… src/SizeBench.AnalysisEngine/DIAInterop/IDiaEnumSymbolsHandCoded.cs
… src/SizeBench.AnalysisEngine/DIAInterop/IDiaSessionExtensions.cs
… src/SizeBench.AnalysisEngine/DiffSession.cs